+++ BitFlyerDotNet.LightningApi/Ohlc.cs
@@ -1,6 +1,6 @@
 ﻿//==============================================================================
-// Copyright (c) 2017-2018 Fiats Inc. All rights reserved.
-// http://www.fiats.asia/
+// Copyright (c) 2017-2019 Fiats Inc. All rights reserved.
+// https://www.fiats.asia/
 //
 
 using System;
@@ -10,30 +10,30 @@ namespace BitFlyerDotNet.LightningApi
     public interface IBfOhlc
     {
         DateTime Start { get; }
-        double Open { get; }
-        double High { get; }
-        double Low { get; }
-        double Close { get; }
-        double Volume { get; }
-        double VWAP { get; }
+        decimal Open { get; }
+        decimal High { get; }
+        decimal Low { get; }
+        decimal Close { get; }
+        decimal Volume { get; }
+        decimal VWAP { get; }
     }
 
     public class BfOhlc : IBfOhlc
     {
         public DateTime Start { get; }
-        public double Open { get; }
-        public double High { get; }
-        public double Low { get; }
-        public double Close { get; }
-        public double Volume { get; }
-        public double VWAP { get; }
+        public decimal Open { get; }
+        public decimal High { get; }
+        public decimal Low { get; }
+        public decimal Close { get; }
+        public decimal Volume { get; }
+        public decimal VWAP { get; }
     }
 
     public interface IBfOhlcEx : IBfOhlc
     {
         int ExecutionCount { get; }
-        double BuyVolume { get; }
-        double SellVolume { get; }
-        double ExecutedVolume { get; }
+        decimal BuyVolume { get; }
+        decimal SellVolume { get; }
+        decimal ExecutedVolume { get; }
     }
 }
