+++ BitFlyerDotNet.LightningApi/Private/GetCollateral.cs
@@ -1,6 +1,6 @@
 ﻿//==============================================================================
-// Copyright (c) 2017-2018 Fiats Inc. All rights reserved.
-// http://www.fiats.asia/
+// Copyright (c) 2017-2019 Fiats Inc. All rights reserved.
+// https://www.fiats.asia/
 //
 
 using Newtonsoft.Json;
@@ -10,16 +10,16 @@ namespace BitFlyerDotNet.LightningApi
     public class BfCollateral
     {
         [JsonProperty(PropertyName = "collateral")]
-        public double Collateral { get; private set; }
+        public decimal Collateral { get; private set; }
 
         [JsonProperty(PropertyName = "open_position_pnl")]
-        public double OpenPositionProfitAndLoss { get; private set; }
+        public decimal OpenPositionProfitAndLoss { get; private set; }
 
         [JsonProperty(PropertyName = "require_collateral")]
-        public double RequireCollateral { get; private set; }
+        public decimal RequireCollateral { get; private set; }
 
         [JsonProperty(PropertyName = "keep_rate")]
-        public double KeepRate { get; private set; }
+        public decimal KeepRate { get; private set; }
     }
 
     public partial class BitFlyerClient
