+++ BitFlyerDotNet.LightningApi/Private/GetPrivateExecutions.cs
@@ -1,6 +1,6 @@
 ﻿//==============================================================================
-// Copyright (c) 2017-2018 Fiats Inc. All rights reserved.
-// http://www.fiats.asia/
+// Copyright (c) 2017-2019 Fiats Inc. All rights reserved.
+// https://www.fiats.asia/
 //
 
 using System;
@@ -23,13 +23,13 @@ namespace BitFlyerDotNet.LightningApi
         public BfTradeSide Side { get; private set; }
 
         [JsonProperty(PropertyName = "price")]
-        public double Price { get; private set; }
+        public decimal Price { get; private set; }
 
         [JsonProperty(PropertyName = "size")]
-        public double Size { get; private set; }
+        public decimal Size { get; private set; }
 
         [JsonProperty(PropertyName = "commission")]
-        public double Commission { get; private set; }
+        public decimal Commission { get; private set; }
 
         [JsonProperty(PropertyName = "exec_date")]
         public DateTime ExecutedTime { get; private set; }
