+++ BitFlyerDotNet.Trading/TradeTicker.cs
@@ -1,23 +1,21 @@
 ﻿//==============================================================================
-// Copyright (c) 2017-2018 Fiats Inc. All rights reserved.
-// http://www.fiats.asia/
+// Copyright (c) 2017-2019 Fiats Inc. All rights reserved.
+// https://www.fiats.asia/
 //
 
 using System;
-using System.Collections.Generic;
-using System.Text;
 using BitFlyerDotNet.LightningApi;
 
 namespace BitFlyerDotNet.Trading
 {
     public class TradeTicker
     {
-        public double Ask { get; set; }
-        public double Bid { get; set; }
-        public double TradePrice { get; set; }
+        public decimal Ask { get; set; }
+        public decimal Bid { get; set; }
+        public decimal TradePrice { get; set; }
         public DateTime UpdatedTime { get; set; }
-        public double SFDDifference { get; set; }
-        public double SFDRate { get; set; }
+        public decimal SFDDifference { get; set; }
+        public decimal SFDRate { get; set; }
         public BfBoardHealth  ServerBusyStatus { get; set; }
     }
 }
